misc/HtmlTestSuite/results/test-1309.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml"><head><link href="mailto:jarhar@chromium.org" rel="author"/>
<link href="https://bugs.chromium.org/p/chromium/issues/detail?id=1300587" rel="help"/>
<script src="/resources/testharness.js"/>
<script src="/resources/testharnessreport.js"/>
<script src="/resources/testdriver.js"/>
<script src="/resources/testdriver-vendor.js"/>

</head><body><form>
  <label for="custom">label</label>
  <my-custom-element id="custom"/>
</form>

<script>
class MyCustomElement extends HTMLElement {
  static formAssociated = true;
  constructor() {
    super();
    const root = this.attachShadow({
      delegatesFocus: true,
      mode: 'open'
    });
    root.appendChild(document.createElement('input'));
  }
};
customElements.define('my-custom-element', MyCustomElement);

window.onload = () =&gt; {
  promise_test(async () =&gt; {
    const label = document.querySelector('label');
    const customElement = document.querySelector('my-custom-element');
    const input = customElement.shadowRoot.querySelector('input');
    await test_driver.click(label);
    assert_equals(document.activeElement, customElement);
    assert_equals(customElement.shadowRoot.activeElement, input);
  }, `Clicking on a label for a form associated custom element with delegatesFocus should focus the custom element's focus delegate.`);
};
</script>
</body></html>
